// src/app/modules/order-dispatch/Models/InvCntsModel.ts
export class InvCntsModel {
    AcceptedInv: number = 0;
    BranchId: number = 0;
    CancelInv: number = 0;
    CompId: number = 0;
    Concern: number = 0;
    Dispatched: number = 0;
    GetpassGenerated: number = 0;
    InvDate: string = '';
    InvoiceDrawn: number = 0;
    Packed: number = 0;
    PendingForAcceptInv: number = 0;
    ReadyToDispatch: number = 0;
    TotalInv: number = 0;
}